Added "nullifyNewStrings"
This is an experimental feature I added because I wanted a way to tell how much of a language has been translated, and this isn't possible when new strings automatically get the same value as the default language. I'm calling the option "nullifyNewStrings". It currently doesn't do anything with dot-notation, since I haven't used that, and I may have overlooked something else as well since I'm not too familiar with this codebase. I'd be glad to get some feedback.